backend-requirements.txt
# Pursuit Response Platform - Backend Dependencies
# Python 3.11+

# Web Framework
fastapi[all]>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
chromadb>=0.4.22
anyio>=4.5.0
mcp==1.22.0

anthropic==0.18.1
openai==1.12.0

# Task Queue
celery[redis]==5.3.4
redis==4.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2

# Document Processing
python-docx==1.1.0
python-pptx==0.6.23
PyPDF2==3.0.1
pdfplumber==0.10.3

# Data Validation
pydantic>=2.5.2
pydantic-settings>=2.1.0
email-validator==2.1.0

# Utilities
structlog==23.3.0
python-dateutil==2.8.2

# Development Dependencies (requirements-dev.txt)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
# black==23.12.0
# isort==5.13.0
# flake8==6.1.0
# mypy==1.7.1
# httpx==0.25.2  # for testing
